--- sekkei.md
@@ -1,7 +1,0 @@
-# React-Sougou-1
-
-ログイン、
-ログインを押したら、データパスワードを照会
-紹介した結果ふたつ合致したら、アドレスパスワードをローカルストレージに保管
-
-とりあえずローカルストレージに保存できた
