+++ src/App.tsx
@@ -1,7 +1,7 @@
 import { useState } from 'react'
 import { ChakraProvider, Box, Flex, Text } from '@chakra-ui/react'
-import HeaderUser from './components/container/HeaderUser/HeaderUser'
-import Search from './components/container/search/Search'
+import HeaderUser from './components/HeaderUser/HeaderUser'
+import Search from './components/search/Search'
 import './App.css'
 import { BrowserRouter, useNavigate } from 'react-router-dom'
 import { UserProvider, useUser } from './Contexts/UserContext'
