+++ src/components/HeaderUser/HeaderUser.tsx
@@ -1,5 +1,5 @@
-import UserMenu from '../../ui/UserMenu/UserMenu'
-import { useHeaderUser } from '../../../hooks/useHeaderUser'
+import UserMenu from '../ui/UserMenu/UserMenu'
+import { useHeaderUser } from '../../hooks/useHeaderUser'
 
 const HeaderUser = () => {
   const { isLoggedIn, handleLogout, handleSignup, handleLogin } = useHeaderUser()
