+++ src/components/ui/AddPinForm/AddPinForm.stories.ts
@@ -24,6 +24,8 @@ const mockFormData = {
   description: '',
   manufacturer: '',
   address: '',
+  lat: 0,
+  lng: 0,
 }
 
 export const Default: Story = {
@@ -44,6 +46,8 @@ export const WithData: Story = {
       description: '飲料',
       manufacturer: 'コカ・コーラ',
       address: '東京都渋谷区渋谷1-1-1',
+      lat: 35.6595,
+      lng: 139.7005,
     },
     isLoadingAddress: false,
     onInputChange: () => {},
@@ -60,6 +64,8 @@ export const RealData: Story = {
       description: '飲料自販機',
       manufacturer: '日本コカ・コーラ株式会社',
       address: '東京都世田谷区学芸大学',
+      lat: 35.6294,
+      lng: 139.6268,
     },
     isLoadingAddress: false,
     onInputChange: () => {},
@@ -76,6 +82,8 @@ export const WithMapData: Story = {
       description: '',
       manufacturer: '',
       address: '駒澤大学',
+      lat: 35.6333,
+      lng: 139.6612,
     },
     isLoadingAddress: false,
     onInputChange: () => {},
