+++ src/data/mockMapPins.ts
@@ -1,0 +1,5 @@
+import type { MapPin } from '../types/pin'
+
+export const mockMapPins: MapPin[] = [
+  { id: 1, lat: 35.6595, lng: 139.7005, address: 'PINの取得が失敗' },
+]
