+++ src/hooks/useEditPinForm/useEditPinForm.ts
@@ -10,7 +10,7 @@ import {
   type PinDetailData 
 } from '../../services/pinDetailService'
 import { geocodeAddress, reverseGeocode } from '../../logic/geocoding/geocodingService'
-import { useCurrentLocation } from '../../logic/centerSpot'
+import { useCurrentLocation } from '../../logic/centerSpot/useCurrentLocation'
 import type { AddPinFormData, Location } from '../../types/addPin'
 
 export const useEditPinForm = () => {
@@ -133,6 +133,14 @@ export const useEditPinForm = () => {
     if (!event.latLng) return
     const lat = event.latLng.lat()
     const lng = event.latLng.lng()
+
+    setSelectedLocation({ lat, lng })
+    setFormData((prev) => ({
+      ...prev,
+      lat,
+      lng,
+    }))
+
     setIsLoadingAddress(true)
     try {
       const address = await reverseGeocode(lat, lng)
@@ -140,10 +148,7 @@ export const useEditPinForm = () => {
         setFormData((prev) => ({
           ...prev,
           address,
-          lat,
-          lng,
         }))
-        setSelectedLocation({ lat, lng })
       }
     } catch (error) {
       console.error('逆ジオコーディングエラー:', error)
