+++ src/pages/AddPinPage.tsx
@@ -1,5 +1,5 @@
 import { Box, VStack, Text } from '@chakra-ui/react'
-import Search from '../components/container/search/Search'
+import Search from '../components/search/Search'
 import { useAddPinForm } from '../hooks/useAddPinForm/useAddPinForm'
 import { AddPinForm } from '../components/ui/AddPinForm/AddPinForm'
 import { MapSelector } from '../components/ui/MapSelector/MapSelector'
