+++ src/router/AppRouter.tsx
@@ -1,11 +1,11 @@
 import { Routes, Route } from 'react-router-dom'
 
 import LoginPage from '../pages/LoginPage'
-import SingUpPage from '../pages/SingUpPage'
+import SignUpPage from '../pages/SignUpPage'
 import UserPage from '../pages/UserPage'
 import AddPinPage from '../pages/AddPinPage'
 import EditPinPage from '../pages/EditPinPage'
-import MyGoogleMap from '../components/container/map/MyGoogleMap'
+import MyGoogleMap from '../components/Map/MyGoogleMap'
 
 type AppRouterProps = {
   search: string
@@ -15,7 +15,7 @@ const AppRouter = ({ search }: AppRouterProps) => {
   return (
     <Routes>
       <Route path="/login" element={<LoginPage />} />
-      <Route path="/signup" element={<SingUpPage />} />
+      <Route path="/signup" element={<SignUpPage />} />
       <Route path="/user" element={<UserPage />} />
       <Route path="/add-pin" element={<AddPinPage />} />
       <Route path="/edit-pin/:pinId" element={<EditPinPage />} />
