+++ src/services/config.ts
@@ -1,0 +1,2 @@
+export const API_BASE_URL = 'http://localhost:3001'
+export const API_URL = `${API_BASE_URL}/api`
