+++ src/services/pinService.ts
@@ -1,17 +1,30 @@
 import type { ApiPin, MapBounds } from '../types/pin'
+import { API_URL } from './config'
 
-// 境界内のピンを取得するAPI関数
-export const fetchPinsInBounds = async (bounds: MapBounds): Promise<ApiPin[]> => {
-  try {
-    const response = await fetch(
-      `http://localhost:3001/api/map_pins?ne_lat=${bounds.ne.lat}&ne_lng=${bounds.ne.lng}&sw_lat=${bounds.sw.lat}&sw_lng=${bounds.sw.lng}`
-    )
-    const pins = await response.json()
-
-    console.log('境界内のピン:', pins)
-    return pins
-  } catch (error) {
-    console.error('ピン取得エラー:', error)
-    throw error
+export const fetchAllMapPins = async (): Promise<ApiPin[]> => {
+  const response = await fetch(`${API_URL}/map_pins`)
+  if (!response.ok) {
+    throw new Error('ピン取得に失敗しました')
   }
+  return response.json()
+}
+
+export const searchMapPins = async (query: string): Promise<ApiPin[]> => {
+  const response = await fetch(
+    `${API_URL}/map_pins/search?query=${encodeURIComponent(query)}`
+  )
+  if (!response.ok) {
+    throw new Error('ピン検索に失敗しました')
+  }
+  return response.json()
+}
+
+export const fetchPinsInBounds = async (bounds: MapBounds): Promise<ApiPin[]> => {
+  const response = await fetch(
+    `${API_URL}/map_pins?ne_lat=${bounds.ne.lat}&ne_lng=${bounds.ne.lng}&sw_lat=${bounds.sw.lat}&sw_lng=${bounds.sw.lng}`
+  )
+  if (!response.ok) {
+    throw new Error('ピン取得に失敗しました')
+  }
+  return response.json()
 }
